LLM Introduction/pyproject.toml
[tool.poetry]
name = "hw_1"
version = "0.0.1"
description = "Homework 1 project with required dependencies"
authors = ["Sergey Yurpalov <sergey.yurpalov@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
gigachain = "0.2.6"
gigachain_community = "0.2.6"
gigachain-cli = "0.0.25"
duckduckgo-search = "6.2.4"
pyfiglet = "1.0.2"
langchain-anthropic = "*"
llama_index = "0.9.40"
pypdf = "4.0.1"
sentence_transformers = "^2.3.1"
python-dotenv = "^1.0.0"
libcst = "1.4.0"
langchain-huggingface = "^0.1.0"
jupyter = "^1.1.1"
requests = "^2.32.3"
pydantic = "^2.9.2"
faiss-cpu = "^1.9.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
mypy = "^1.11.2"
flake8 = "^7.1.1"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::Warning"
]

[tool.mypy]
mypy_path = "hw_1"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
